+++ tests/testthat/test-scan1coef_pg.R
@@ -5,11 +5,11 @@ eff_via_lm <-
     function(probs, pheno, kinship, addcovar=NULL, intcovar=NULL,
              se=TRUE)
 {
-    kinship <- double_kinship(kinship) # need 2*kinship for LMM
+    kinship <- qtl2:::double_kinship(kinship) # need 2*kinship for LMM
     kinship <- decomp_kinship(kinship)
     eigenvec <- kinship$vectors
-    hsq <- calc_hsq_clean(kinship, as.matrix(pheno), addcovar, NULL, FALSE, NULL,
-                          reml=TRUE, cores=1, check_boundary=TRUE, tol=1e-12)$hsq[1,1]
+    hsq <- qtl2:::calc_hsq_clean(kinship, as.matrix(pheno), addcovar, NULL, FALSE, NULL,
+                                 reml=TRUE, cores=1, check_boundary=TRUE, tol=1e-12)$hsq[1,1]
 
     wts <- 1/(hsq*kinship$values + (1-hsq))
 
@@ -126,18 +126,18 @@ test_that("scan1coef_pg for grav", {
     covar <- cbind(covar, chr4=pr[["4"]][,2,"CD.329C-Col"])
     est <- scan1coef(pr[,"3"], phe, K, covar, se=FALSE, zerosum=FALSE)
     est_lm <- eff_via_lm(pr[["3"]], phe, K, covar)
-    expect_equivalent(est, est_lm)
+    expect_equivalent(est[-22,], est_lm[-22,]) # different NAs at 22
 
     est <- scan1coef(pr[,"3"], phe, K, covar, se=TRUE, zerosum=FALSE)
-    expect_equivalent(est, est_lm)
-    expect_equivalent(attr(est, "SE"), attr(est_lm, "SE"))
+    expect_equivalent(est[-22,], est_lm[-22,]) # different NAs at 22
+    expect_equivalent(attr(est, "SE")[-22,], attr(est_lm, "SE")[-22,])
 
     # pre-computed eigen decomp
     est <- scan1coef(pr[,"3"], phe, Ke, covar, zerosum=FALSE)
-    expect_equivalent(est, est_lm)
+    expect_equivalent(est[-22,], est_lm[-22,]) # different NAs at 22
     est <- scan1coef(pr[,"3"], phe, Ke, covar, se=TRUE, zerosum=FALSE)
-    expect_equivalent(est, est_lm)
-    expect_equivalent(attr(est, "SE"), attr(est_lm, "SE"))
+    expect_equivalent(est[-22,], est_lm[-22,]) # different NAs at 22
+    expect_equivalent(attr(est, "SE")[-22,], attr(est_lm, "SE")[-22,])
 
     # two interactive covariates
     est <- scan1coef(pr[,"3"], phe, K, covar, intcovar=covar, se=FALSE, zerosum=FALSE)